feat(sheet): move the merge-cell logic from row/col-mutation to merge-cell-controller
