+++ .vscode/settings.json
@@ -8,6 +8,7 @@
     "commentTranslate.hover.enabled": true,
     "commentTranslate.source": "Google",
     "cSpell.words": [
+        "undos",
         "Bandings",
         "Catmull",
         "clazz",
